feat: impl. Metal & Dielectric

+++ cuda/include/material.cuh
@@ -80,7 +80,7 @@ public:
 		double refraction_ratio = rec.front_face ? (1.0 / ir) : ir;
 
 		vec3 unit_direction = unit_vector(r_in.direction);
-		double cos_theta = fmin(dot(-unit_direction, rec.normal), 1.0);
+		double cos_theta = fminf(dot(-unit_direction, rec.normal), 1.0);
 		double sin_theta = sqrt(1.0 - cos_theta * cos_theta);
 
 		bool cannot_refract = refraction_ratio * sin_theta > 1.0;
