+++ cuda/include/vec3.cuh
@@ -63,8 +63,8 @@ __device__ inline vec3 random_unit_vector(curandState* randState) {
 }
 
 
-__device__ inline vec3 refract(vec3 uv, const vec3& n, double etai_over_etat) {
-	auto cos_theta = fmin(dot(-uv, n), 1.0);
+__device__ inline vec3 refract(vec3 uv, const vec3& n, float etai_over_etat) {
+	auto cos_theta = fminf(dot(-uv, n), 1.0);
 	vec3 r_out_perp = etai_over_etat * (uv + cos_theta * n);
 	vec3 r_out_parallel = -sqrt(fabs(1.0 - length_squared(r_out_perp))) * n;
 	return r_out_perp + r_out_parallel;
